src/requirements.txt
h5py
numpy
Pillow
